feat: Add axios dependency and configure proxy for API requests
This commit adds the axios dependency to the project, which will be used for making HTTP requests. It also configures a proxy in the vite.config.js file to redirect requests to the '/api' path to the 'https://ctftime.org' URL. This will allow the application to make API requests to ctftime.org without running into cross-origin issues.

Ref: #123